+++ Test/Integration/Service/EmailSenderTest.php
@@ -19,46 +19,129 @@ namespace MultiSafepay\ConnectCore\Test\Integration\Service;
 
 use Exception;
 use Magento\Framework\App\Config\ScopeConfigInterface;
+use Magento\Framework\Exception\LocalizedException;
 use Magento\Sales\Api\Data\InvoiceInterface;
 use Magento\Sales\Api\Data\OrderPaymentInterface;
+use Magento\Sales\Model\Order\Email\Container\OrderIdentity;
 use Magento\Sales\Model\Order\Email\Sender\InvoiceSender;
 use Magento\Sales\Model\Order\Email\Sender\OrderSender;
+use MultiSafepay\ConnectCore\Model\Ui\Gateway\BankTransferConfigProvider;
 use MultiSafepay\ConnectCore\Service\EmailSender;
 use MultiSafepay\ConnectCore\Test\Integration\AbstractTestCase;
 
 class EmailSenderTest extends AbstractTestCase
 {
     /**
+     * @magentoDataFixture   Magento/Sales/_files/order.php
+     *
+     * @throws LocalizedException
+     * @throws Exception
+     */
+    public function testsendOrderConfirmationEmailWithEmailAlreadySent(): void
+    {
+        $this->getOrder()->setEmailSent(1);
+
+        $emailSender = $this->getEmailSender()->sendOrderConfirmationEmail(
+            $this->getOrder(),
+            EmailSender::AFTER_TRANSACTION_EMAIL_TYPE
+        );
+
+        self::assertFalse($emailSender);
+    }
+
+    /**
+     * @magentoDataFixture   Magento/Sales/_files/order.php
+     *
+     * @throws LocalizedException
+     * @throws Exception
+     */
+    public function testsendOrderConfirmationEmailWithEmailSendingOff(): void
+    {
+        $this->getOrder()->setEmailSent(0);
+
+        $emailSender = $this->getEmailSender(false)->sendOrderConfirmationEmail(
+            $this->getOrder(),
+            EmailSender::AFTER_TRANSACTION_EMAIL_TYPE
+        );
+
+        self::assertFalse($emailSender);
+    }
+
+    /**
+     * @magentoDataFixture   Magento/Sales/_files/order.php
+     *
+     * @throws LocalizedException
+     * @throws Exception
+     */
+    public function testsendOrderConfirmationEmailWithEmailSendingOn(): void
+    {
+        $this->getOrder()->setEmailSent(0);
+
+        $emailSender = $this->getEmailSender(true)->sendOrderConfirmationEmail(
+            $this->getOrder(),
+            EmailSender::AFTER_TRANSACTION_EMAIL_TYPE
+        );
+
+        self::assertTrue($emailSender);
+    }
+
+    /**
+     * @magentoConfigFixture default_store multisafepay/general/order_confirmation_email after_transaction
+     */
+    public function testcheckOrderConfirmationBeforeTransactionWithEmptyGatewaySpecificSettingIsFalse(): void
+    {
+        self::assertFalse(
+            $this->getEmailSender()->checkOrderConfirmationBeforeTransaction(BankTransferConfigProvider::CODE)
+        );
+    }
+
+    /**
+     * @magentoConfigFixture default_store multisafepay/general/order_confirmation_email before_transaction
+     */
+    public function testcheckOrderConfirmationBeforeTransactionWithEmptyGatewaySpecificSettingIsTrue(): void
+    {
+        self::assertTrue(
+            $this->getEmailSender()->checkOrderConfirmationBeforeTransaction(BankTransferConfigProvider::CODE)
+        );
+    }
+
+    /**
+     * @magentoConfigFixture default_store multisafepay/general/order_confirmation_email after_transaction
+     * @magentoConfigFixture default_store payment/multisafepay_banktransfer/order_confirmation_email after_transaction
+     */
+    public function testcheckOrderConfirmationBeforeTransactionWithGatewaySpecificSettingIsFalse(): void
+    {
+        self::assertFalse(
+            $this->getEmailSender()->checkOrderConfirmationBeforeTransaction(BankTransferConfigProvider::CODE)
+        );
+    }
+
+    /**
+     * @magentoConfigFixture default_store multisafepay/general/order_confirmation_email after_transaction
+     * @magentoConfigFixture default_store payment/multisafepay_banktransfer/order_confirmation_email before_transaction
+     */
+    public function testcheckOrderConfirmationBeforeTransactionWithGatewaySpecificSettingIsTrue(): void
+    {
+        self::assertTrue(
+            $this->getEmailSender()->checkOrderConfirmationBeforeTransaction(BankTransferConfigProvider::CODE)
+        );
+    }
+
+    /**
      * Test to see if an invoice could be sent
      *
      * @magentoConfigFixture default_store sales_email/invoice/enabled 1
      * @throws Exception
      */
-    public function testSendInvoiceEmailWhileEmailHasBeenSentAlready()
+    public function testSendInvoiceEmailWhileEmailHasBeenSentAlready(): void
     {
         $orderPayment = $this->getNewOrderPayment();
         $invoice = $this->getNewInvoice();
         $invoice->setEmailSent(true);
 
         $emailSender = $this->getEmailSender();
-        $this->assertFalse($emailSender->sendInvoiceEmail($orderPayment, $invoice));
+        self::assertFalse($emailSender->sendInvoiceEmail($orderPayment, $invoice));
     }
-
-    /**
-     * Test to see if an invoice could be sent
-     *
-     * @magentoConfigFixture default_store sales_email/invoice/enabled foobar
-     * @throws Exception
-     */
-    /*public function testSendInvoiceEmailWhenEmailingHasBeenDisabled()
-    {
-        $this->assertEquals(0, $this->getScopeConfig()->getValue('sales_email/invoice/enabled'));
-
-        $orderPayment = $this->getNewOrderPayment();
-        $invoice = $this->getNewInvoice();
-        $emailSender = $this->getEmailSender();
-        $this->assertFalse($emailSender->sendInvoiceEmail($orderPayment, $invoice));
-    }*/
 
     /**
      * Test to see if an invoice could be sent
@@ -66,24 +149,33 @@ class EmailSenderTest extends AbstractTestCase
      * @magentoConfigFixture default_store sales_email/invoice/enabled 1
      * @throws Exception
      */
-    public function testSendInvoiceEmailWhileAllGoesWell()
+    public function testSendInvoiceEmailWhileAllGoesWell(): void
     {
-        $this->assertEquals(1, $this->getScopeConfig()->getValue('sales_email/invoice/enabled'));
+        self::assertEquals(1, $this->getScopeConfig()->getValue('sales_email/invoice/enabled'));
 
         $orderPayment = $this->getNewOrderPayment();
         $invoice = $this->getNewInvoice();
         $emailSender = $this->getEmailSender();
-        $this->assertTrue($emailSender->sendInvoiceEmail($orderPayment, $invoice));
+        self::assertTrue($emailSender->sendInvoiceEmail($orderPayment, $invoice));
     }
 
     /**
+     * @param bool $isEnabled
      * @return EmailSender
      */
-    private function getEmailSender(): EmailSender
+    private function getEmailSender(bool $isEnabled = null): EmailSender
     {
+        $orderIdentityMock = $this->getMockBuilder(OrderIdentity::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $orderIdentityMock->method('isEnabled')->willReturn($isEnabled);
+
         $arguments = [];
         $arguments['invoiceSender'] = $this->createMock(InvoiceSender::class);
         $arguments['orderSender'] = $this->createMock(OrderSender::class);
+        $arguments['orderIdentity'] = $orderIdentityMock;
+
         return $this->getObjectManager()->create(EmailSender::class, $arguments);
     }